add comments and cleanup App.js
