+++ examples/with-one-login-many-subdomains/frontend/src/utils.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import Session from "supertokens-auth-react/recipe/session";
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
 export const websitePort = process.env.REACT_APP_WEBSITE_PORT || 3000;
@@ -8,7 +9,7 @@ export function getApiDomain() {
   return apiUrl;
 }
 
-export const userSubdomainApi = `${getApiDomain()}/user-subdomain`
+export const userSubdomainApi = `${getApiDomain()}/user-subdomain`;
 
 export function getAuthDomain() {
   const websiteUrl =
@@ -19,15 +20,27 @@ export function getAuthDomain() {
 
 export async function getSubdomainForCurrentUser() {
   const res = await axios.get(userSubdomainApi);
-  return res.data.subdomain
+  return res.data.subdomain;
 }
 
 export async function getRedirectionUrlForUser() {
-  try {
-    const subdomain = await getSubdomainForCurrentUser()
-    return `http://${subdomain}.example.com:${websitePort}`;
-  } catch (error) {
-    return getAuthDomain();
-  }
+  const subdomain = await getSubdomainForCurrentUser();
+  return `http://${subdomain}.example.com:${websitePort}`;
 }
 
+export async function redirectIfOnWrongSubdomain() {
+  try {
+    if (Session.doesSessionExist()) {
+      const currentSubdomain = window.location.hostname.split(".")[0];
+      const currentUserSubdomain = await getSubdomainForCurrentUser();
+      // location.origin check ensures that user gets the option to click
+      // the continue button on verify-email page
+      if (
+        window.location.origin !== getAuthDomain() &&
+        currentSubdomain !== currentUserSubdomain
+      ) {
+        window.location.href = `http://${currentUserSubdomain}.example.com:${websitePort}`;
+      }
+    }
+  } catch (error) {}
+}
